requirements.txt
haystack-ai==2.18.1
chroma-haystack==3.3.0
sentence-transformers==5.1.1
llama-cpp-python==0.3.16
llama-cpp-haystack==1.3.0
fastapi==0.119.0
uvicorn==0.37.0
python-multipart==0.0.20
